featurizer/requirements.txt
# Core FastAPI dependencies
fastapi==0.123.5
uvicorn[standard]==0.33.0
pydantic==2.12.5
httpx==0.28.1

# Eureka service discovery
py-eureka-client==0.13.2

# Database
psycopg2-binary==2.9.11
sqlalchemy==2.1.4
alembic==1.17.2

# Data processing (lighter versions)
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2

# NLP - lightweight alternatives
# Use lighter spacy model or remove if not critical
# spacy>=3.7.0,<3.8.0
# en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Text processing (much lighter than sentence-transformers)
regex==2024.11.6

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
requests==2.32.5
PyYAML==6.0.2
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
